requirements.txt
Django==5.1.2
pillow==10.4.0
psycopg2==2.9.10
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
drf_yasg==1.21.8




